update cumulative time to an int

+++ src/main/java/com/kendelong/util/performance/PerformanceMonitor.java
@@ -91,8 +91,8 @@ public class PerformanceMonitor
 		return numberOfExceptions.get();
 	}
 	
-	public double getCumulativeTime()
+	public int getCumulativeTime()
 	{
-		return getAverageResponseTime()*getNumberOfAccesses();
+		return (int) (getAverageResponseTime()*getNumberOfAccesses());
 	}
 }
